Added support for excluding files based on file system patterns (e.g. "*.html") via the new '-x' switch.
File exclusion is powered by the fnmatch() function. In MinGW, libiberty.a
(under MSYS) has the implementation, but there is no header file that declares
it so the function declaration was added to the top of main.cpp for now.
This is not ideal and it has not been tested in any other platforms yet.